+++ src/app/expandablearea/expandablearea-footer.demo.ts
@@ -1,0 +1,15 @@
+import {
+  Component,
+  OnInit
+} from '@angular/core';
+
+@Component({
+  selector: 'soho-expandablearea-footer-demo',
+  templateUrl: './expandablearea-footer.demo.html'
+})
+export class ExpandableAreaFooterDemoComponent implements OnInit {
+
+  constructor() {}
+
+  ngOnInit() { }
+}